split and simplify route routes
